Start splitting object index from object manager file
