+++ src/OpenLoco/Objects/ObjectIndex.cpp
@@ -1,0 +1,585 @@
+#include "../Environment.h"
+#include "../Interop/Interop.hpp"
+#include "../Localisation/StringIds.h"
+#include "../Localisation/StringManager.h"
+#include "../S5/SawyerStream.h"
+#include "../Ui.h"
+#include "../Ui/ProgressBar.h"
+#include "../Utility/Numeric.hpp"
+#include "../Utility/Stream.hpp"
+#include "ObjectManager.h"
+#include <cstdint>
+#include <fstream>
+
+using namespace OpenLoco::Interop;
+
+namespace OpenLoco::ObjectManager
+{
+    static loco_global<std::byte*, 0x0050D13C> _installedObjectList;
+    static loco_global<uint32_t, 0x0112A110> _installedObjectCount;
+    static loco_global<bool, 0x0112A17E> _customObjectsInIndex;
+    static loco_global<uint32_t, 0x0050D154> _totalNumImages;
+    static loco_global<std::byte*, 0x0050D158> _50D158;
+    static loco_global<std::byte[0x2002], 0x0112A17F> _112A17F;
+    static loco_global<bool, 0x0050AEAD> _isFirstTime;
+    static loco_global<bool, 0x0050D161> _isPartialLoaded;
+    static loco_global<uint32_t, 0x009D9D52> _decodedSize;    // return of getScenarioText (badly named)
+    static loco_global<uint32_t, 0x0112A168> _numImages;      // return of getScenarioText (badly named)
+    static loco_global<uint8_t, 0x0112C211> _intelligence;    // return of getScenarioText (badly named)
+    static loco_global<uint8_t, 0x0112C212> _aggressiveness;  // return of getScenarioText (badly named)
+    static loco_global<uint8_t, 0x0112C213> _competitiveness; // return of getScenarioText (badly named)
+
+#pragma pack(push, 1)
+    struct ObjectFolderState
+    {
+        uint32_t numObjects = 0;
+        uint32_t totalFileSize = 0;
+        uint32_t dateHash = 0;
+        constexpr bool operator==(const ObjectFolderState& rhs) const
+        {
+            return (numObjects == rhs.numObjects) && (totalFileSize == rhs.totalFileSize) && (dateHash == rhs.dateHash);
+        }
+        constexpr bool operator!=(const ObjectFolderState& rhs) const
+        {
+            return !(*this == rhs);
+        }
+    };
+    static_assert(sizeof(ObjectFolderState) == 0xC);
+    struct IndexHeader
+    {
+        ObjectFolderState state;
+        uint32_t fileSize;
+        uint32_t numObjects; // duplicates ObjectFolderState.numObjects but without high 1 and includes corrupted .dat's
+    };
+    static_assert(sizeof(IndexHeader) == 0x14);
+#pragma pack(pop)
+
+    // 0x00470F3C
+    static ObjectFolderState getCurrentObjectFolderState()
+    {
+        ObjectFolderState currentState;
+        const auto objectPath = Environment::getPathNoWarning(Environment::PathId::objects);
+        for (const auto& file : fs::directory_iterator(objectPath, fs::directory_options::skip_permission_denied))
+        {
+            if (!file.is_regular_file())
+            {
+                continue;
+            }
+            const auto extension = file.path().extension().u8string();
+            if (!Utility::iequals(extension, ".DAT"))
+            {
+                continue;
+            }
+            currentState.numObjects++;
+            const auto lastWrite = file.last_write_time().time_since_epoch().count();
+            currentState.dateHash ^= ((lastWrite >> 32) ^ (lastWrite & 0xFFFFFFFF));
+            currentState.dateHash = Utility::ror(currentState.dateHash, 5);
+            currentState.totalFileSize += file.file_size();
+        }
+        currentState.numObjects |= (1 << 24);
+        return currentState;
+    }
+
+    // 0x00471712
+    static bool hasCustomObjectsInIndex()
+    {
+        auto* ptr = *_installedObjectList;
+        for (uint32_t i = 0; i < _installedObjectCount; i++)
+        {
+            auto entry = ObjectIndexEntry::read(&ptr);
+            if (entry._header->isCustom())
+                return true;
+        }
+        return false;
+    }
+
+    static void saveIndex(const IndexHeader& header)
+    {
+        std::ofstream stream;
+        const auto indexPath = Environment::getPathNoWarning(Environment::PathId::plugin1);
+        stream.open(indexPath, std::ios::out | std::ios::binary);
+        if (!stream.is_open())
+        {
+            return;
+        }
+        stream.write(reinterpret_cast<const char*>(&header), sizeof(header));
+        stream.write(reinterpret_cast<const char*>(*_installedObjectList), header.fileSize);
+    }
+
+    static std::pair<ObjectIndexEntry, size_t> createPartialNewEntry(std::byte* entryBuffer, const ObjectHeader& objHeader, const fs::path filename)
+    {
+        ObjectIndexEntry entry{};
+        size_t newEntrySize = 0;
+
+        // Header
+        std::memcpy(&entryBuffer[newEntrySize], &objHeader, sizeof(objHeader));
+        entry._header = reinterpret_cast<ObjectHeader*>(&entryBuffer[newEntrySize]);
+        newEntrySize += sizeof(objHeader);
+
+        // Filename
+        std::strcpy(reinterpret_cast<char*>(&entryBuffer[newEntrySize]), filename.u8string().c_str());
+        entry._filename = reinterpret_cast<char*>(&entryBuffer[newEntrySize]);
+        newEntrySize += strlen(reinterpret_cast<char*>(&entryBuffer[newEntrySize])) + 1;
+
+        // Header2 NULL
+        ObjectHeader2 objHeader2;
+        objHeader2.var_00 = std::numeric_limits<uint32_t>::max();
+        std::memcpy(&entryBuffer[newEntrySize], &objHeader2, sizeof(objHeader2));
+        newEntrySize += sizeof(objHeader2);
+
+        // Name NULL
+        entryBuffer[newEntrySize] = std::byte(0);
+        entry._name = reinterpret_cast<char*>(&entryBuffer[newEntrySize]);
+        newEntrySize += strlen(reinterpret_cast<char*>(&entryBuffer[newEntrySize])) + 1;
+
+        // Header3 NULL
+        ObjectHeader3 objHeader3{};
+        std::memcpy(&entryBuffer[newEntrySize], &objHeader3, sizeof(objHeader3));
+        newEntrySize += sizeof(objHeader3);
+
+        // ObjectList1
+        const uint8_t size1 = 0;
+        entryBuffer[newEntrySize++] = std::byte(size1);
+
+        // ObjectList2
+        const uint8_t size2 = 0;
+        entryBuffer[newEntrySize++] = std::byte(size2);
+
+        return std::make_pair(entry, newEntrySize);
+    }
+
+    static std::pair<ObjectIndexEntry, size_t> createNewEntry(std::byte* entryBuffer, const ObjectHeader& objHeader, const fs::path filename)
+    {
+        ObjectIndexEntry entry{};
+        size_t newEntrySize = 0;
+
+        // Header
+        std::memcpy(&entryBuffer[newEntrySize], &objHeader, sizeof(objHeader));
+        entry._header = reinterpret_cast<ObjectHeader*>(&entryBuffer[newEntrySize]);
+        newEntrySize += sizeof(objHeader);
+
+        // Filename
+        std::strcpy(reinterpret_cast<char*>(&entryBuffer[newEntrySize]), filename.u8string().c_str());
+        entry._filename = reinterpret_cast<char*>(&entryBuffer[newEntrySize]);
+        newEntrySize += strlen(reinterpret_cast<char*>(&entryBuffer[newEntrySize])) + 1;
+
+        // Header2
+        ObjectHeader2 objHeader2;
+        objHeader2.var_00 = _decodedSize;
+        std::memcpy(&entryBuffer[newEntrySize], &objHeader2, sizeof(objHeader2));
+        newEntrySize += sizeof(objHeader2);
+
+        // Name
+        strcpy(reinterpret_cast<char*>(&entryBuffer[newEntrySize]), StringManager::getString(0x2000));
+        entry._name = reinterpret_cast<char*>(&entryBuffer[newEntrySize]);
+        newEntrySize += strlen(reinterpret_cast<char*>(&entryBuffer[newEntrySize])) + 1;
+
+        // Header3
+        ObjectHeader3 objHeader3{};
+        objHeader3.var_00 = _numImages;
+        objHeader3.intelligence = _intelligence;
+        objHeader3.aggressiveness = _aggressiveness;
+        objHeader3.competitiveness = _competitiveness;
+        std::memcpy(&entryBuffer[newEntrySize], &objHeader3, sizeof(objHeader3));
+        newEntrySize += sizeof(objHeader3);
+
+        auto* ptr = &_112A17F[0];
+
+        // ObjectList1 (required objects)
+        const uint8_t size1 = uint8_t(*ptr++);
+        entryBuffer[newEntrySize++] = std::byte(size1);
+        std::memcpy(&entryBuffer[newEntrySize], ptr, sizeof(ObjectHeader) * size1);
+        newEntrySize += sizeof(ObjectHeader) * size1;
+        ptr += sizeof(ObjectHeader) * size1;
+
+        // ObjectList2 (also loads objects {used for category selection})
+        const uint8_t size2 = uint8_t(*ptr++);
+        entryBuffer[newEntrySize++] = std::byte(size2);
+        std::memcpy(&entryBuffer[newEntrySize], ptr, sizeof(ObjectHeader) * size2);
+        newEntrySize += sizeof(ObjectHeader) * size2;
+        ptr += sizeof(ObjectHeader) * size2;
+
+        return std::make_pair(entry, newEntrySize);
+    }
+
+    // Adds a new object to the index by: 1. creating a partial index, 2. validating, 3. creating a full index entry
+    static void addObjectToIndex(const fs::path filepath, size_t& usedBufferSize)
+    {
+        ObjectHeader objHeader{};
+        {
+            std::ifstream stream;
+            stream.open(filepath, std::ios::in | std::ios::binary);
+            Utility::readData(stream, objHeader);
+            if (stream.gcount() != sizeof(objHeader))
+            {
+                return;
+            }
+        }
+
+        const auto curObjPos = usedBufferSize;
+        const auto partialNewEntry = createPartialNewEntry(&_installedObjectList[usedBufferSize], objHeader, filepath.filename());
+        usedBufferSize += partialNewEntry.second;
+        _installedObjectCount++;
+
+        _isPartialLoaded = true;
+        _50D158 = _112A17F;
+        getScenarioText(objHeader);
+        _50D158 = reinterpret_cast<std::byte*>(-1);
+        _isPartialLoaded = false;
+        if (_installedObjectCount == 0)
+        {
+            return;
+        }
+        _installedObjectCount--;
+
+        // Rewind as it is only a partial object loaded
+        usedBufferSize = curObjPos;
+
+        // Load full entry into temp buffer.
+        // 0x009D1CC8
+        std::byte newEntryBuffer[0x2000] = {};
+        const auto [newEntry, newEntrySize] = createNewEntry(newEntryBuffer, objHeader, filepath.filename());
+
+        freeScenarioText();
+
+        auto* indexPtr = *_installedObjectList;
+        // This ptr will be pointing at where in the object list to install the new entry
+        auto* installPtr = indexPtr;
+        for (uint32_t i = 0; i < _installedObjectCount; i++)
+        {
+            auto entry = ObjectIndexEntry::read(&indexPtr);
+            if (strcmp(newEntry._name, entry._name) < 0)
+            {
+                break;
+            }
+            // If cmp < 0 then we will want to install at the previous indexPtr location
+            installPtr = indexPtr;
+        }
+
+        auto moveSize = usedBufferSize - (installPtr - *_installedObjectList);
+        std::memmove(installPtr + newEntrySize, installPtr, moveSize);
+        std::memcpy(installPtr, newEntryBuffer, newEntrySize);
+        usedBufferSize += newEntrySize;
+
+        _installedObjectCount++;
+    }
+
+    // 0x0047118B
+    static void createIndex(const ObjectFolderState& currentState)
+    {
+        Ui::processMessagesMini();
+        const auto progressString = _isFirstTime ? StringIds::starting_for_the_first_time : StringIds::checking_object_files;
+        Ui::ProgressBar::begin(progressString);
+
+        // Reset
+        reloadAll();
+        if (reinterpret_cast<int32_t>(*_installedObjectList) != -1)
+        {
+            free(*_installedObjectList);
+        }
+
+        // Prepare initial object list buffer (we will grow this as required)
+        size_t bufferSize = 0x4000;
+        _installedObjectList = static_cast<std::byte*>(malloc(bufferSize));
+        if (_installedObjectList == nullptr)
+        {
+            _installedObjectList = reinterpret_cast<std::byte*>(-1);
+            exitWithError(StringIds::unable_to_allocate_enough_memory, StringIds::game_init_failure);
+            return;
+        }
+
+        _installedObjectCount = 0;
+        // Create new index by iterating all DAT files and processing
+        IndexHeader header{};
+        uint8_t progress = 0;      // Progress is used for the ProgressBar Ui element
+        size_t usedBufferSize = 0; // Keep track of used space to allow for growth and for final sizing
+        const auto objectPath = Environment::getPathNoWarning(Environment::PathId::objects);
+        for (const auto& file : fs::directory_iterator(objectPath, fs::directory_options::skip_permission_denied))
+        {
+            if (!file.is_regular_file())
+            {
+                continue;
+            }
+            const auto extension = file.path().extension().u8string();
+            if (!Utility::iequals(extension, ".DAT"))
+            {
+                continue;
+            }
+            Ui::processMessagesMini();
+            header.state.numObjects++;
+
+            // Cheap calculation of (curObjectCount / totalObjectCount) * 256
+            const auto newProgress = (header.state.numObjects << 8) / ((currentState.numObjects & 0xFFFFFF) + 1);
+            if (progress != newProgress)
+            {
+                progress = newProgress;
+                Ui::ProgressBar::setProgress(newProgress);
+            }
+
+            // Grow object list buffer if near limit
+            const auto remainingBuffer = bufferSize - usedBufferSize;
+            if (remainingBuffer < 0x231E)
+            {
+                // Original grew buffer at slower rate. Memory is cheap though
+                bufferSize *= 2;
+                _installedObjectList = static_cast<std::byte*>(realloc(*_installedObjectList, bufferSize));
+                if (_installedObjectList == nullptr)
+                {
+                    exitWithError(StringIds::unable_to_allocate_enough_memory, StringIds::game_init_failure);
+                    return;
+                }
+            }
+
+            addObjectToIndex(file.path(), usedBufferSize);
+        }
+
+        // New index creation completed. Reset and save result.
+        reloadAll();
+        header.fileSize = usedBufferSize;
+        header.numObjects = _installedObjectCount;
+        header.state = currentState;
+        saveIndex(header);
+
+        Ui::ProgressBar::end();
+    }
+
+    static bool tryLoadIndex(const ObjectFolderState& currentState)
+    {
+        const auto indexPath = Environment::getPathNoWarning(Environment::PathId::plugin1);
+        if (!fs::exists(indexPath))
+        {
+            return false;
+        }
+        std::ifstream stream;
+        stream.open(indexPath, std::ios::in | std::ios::binary);
+        if (!stream.is_open())
+        {
+            return false;
+        }
+        // 0x00112A14C -> 160
+        IndexHeader header{};
+        Utility::readData(stream, header);
+        if ((header.state != currentState) || (stream.gcount() != sizeof(header)))
+        {
+            return false;
+        }
+        else
+        {
+            if (reinterpret_cast<int32_t>(*_installedObjectList) != -1)
+            {
+                free(*_installedObjectList);
+            }
+            _installedObjectList = static_cast<std::byte*>(malloc(header.fileSize));
+            if (_installedObjectList == nullptr)
+            {
+                exitWithError(StringIds::unable_to_allocate_enough_memory, StringIds::game_init_failure);
+                return false;
+            }
+            Utility::readData(stream, *_installedObjectList, header.fileSize);
+            if (stream.gcount() != static_cast<int32_t>(header.fileSize))
+            {
+                return false;
+            }
+            _installedObjectCount = header.numObjects;
+            reloadAll();
+        }
+        return true;
+    }
+
+    // 0x00470F3C
+    void loadIndex()
+    {
+        // 0x00112A138 -> 144
+        const auto currentState = getCurrentObjectFolderState();
+
+        if (!tryLoadIndex(currentState))
+        {
+            createIndex(currentState);
+        }
+
+        _customObjectsInIndex = hasCustomObjectsInIndex();
+    }
+
+    uint32_t getNumInstalledObjects()
+    {
+        return *_installedObjectCount;
+    }
+
+    std::vector<std::pair<uint32_t, ObjectIndexEntry>> getAvailableObjects(ObjectType type)
+    {
+        auto ptr = (std::byte*)_installedObjectList;
+        std::vector<std::pair<uint32_t, ObjectIndexEntry>> list;
+
+        for (uint32_t i = 0; i < _installedObjectCount; i++)
+        {
+            auto entry = ObjectIndexEntry::read(&ptr);
+            if (entry._header->getType() == type)
+                list.emplace_back(std::pair<uint32_t, ObjectIndexEntry>(i, entry));
+        }
+
+        return list;
+    }
+
+    // 0x0047285C
+    static bool installObject(const ObjectHeader& objectHeader)
+    {
+        // Prepare progress bar
+        char caption[512];
+        auto* str = StringManager::formatString(caption, sizeof(caption), StringIds::installing_new_data);
+        // Convert object name to string so it is properly terminated
+        std::string objectname(objectHeader.getName());
+        strcat(str, objectname.c_str());
+        Ui::ProgressBar::begin(caption);
+        Ui::ProgressBar::setProgress(50);
+        Ui::processMessagesMini();
+
+        // Get new file path
+        std::string filename = objectname;
+        sanatiseObjectFilename(filename);
+        auto objPath = findObjectPath(filename);
+
+        // Create new file and output object file
+        Ui::ProgressBar::setProgress(180);
+        SawyerStreamWriter stream(objPath);
+        writePackedObjects(stream, { objectHeader });
+
+        // Free file
+        stream.close();
+        Ui::ProgressBar::setProgress(240);
+        Ui::ProgressBar::setProgress(255);
+        Ui::ProgressBar::end();
+        return true;
+    }
+
+    static bool isObjectInstalled(const ObjectHeader& objectHeader)
+    {
+        const auto objects = getAvailableObjects(objectHeader.getType());
+        auto res = std::find_if(std::begin(objects), std::end(objects), [&objectHeader](auto& obj) { return *obj.second._header == objectHeader; });
+        return res != std::end(objects);
+    }
+
+    static void permutateObjectFilename(std::string& filename)
+    {
+        auto* firstChar = filename.c_str();
+        auto* endChar = &filename[filename.size()];
+        auto* c = endChar;
+        do
+        {
+            c--;
+            if (c == firstChar)
+            {
+                filename = "00000000";
+                break;
+            }
+            if (*c < '0')
+            {
+                *c = '/';
+            }
+            if (*c == '9')
+            {
+                *c = '@';
+            }
+            if (*c == 'Z')
+            {
+                *c = '/';
+            }
+            *c = *c + 1;
+        } while (*c == '0');
+    }
+
+    static void sanatiseObjectFilename(std::string& filename)
+    {
+        // Trim string at first space (note this copies vanilla but maybe shouldn't)
+        auto space = filename.find_first_of(' ');
+        if (space != std::string::npos)
+        {
+            filename = filename.substr(0, space);
+        }
+        // Make filename uppercase
+        std::transform(std::begin(filename), std::end(filename), std::begin(filename), toupper);
+    }
+
+    // All object files are based on their internal object header name but
+    // there is a chance of a name collision this function works out if the name
+    // is possible and if not permutates the name until it is valid.
+    static fs::path findObjectPath(std::string& filename)
+    {
+        auto objPath = Environment::getPath(Environment::PathId::objects);
+
+        bool permutateName = false;
+        do
+        {
+            if (permutateName)
+            {
+                permutateObjectFilename(filename);
+            }
+            objPath.replace_filename(filename);
+            objPath.replace_extension(".DAT");
+            permutateName = true;
+        } while (fs::exists(objPath));
+        return objPath;
+    }
+
+    // 0x00472687 based on
+    bool tryInstallObject(const ObjectHeader& objectHeader, stdx::span<const uint8_t> data)
+    {
+        unloadAll();
+        if (!computeObjectChecksum(objectHeader, data))
+        {
+            return false;
+        }
+        // Copy the object into Loco freeable memory (required for when partialLoad loads the object)
+        uint8_t* objectData = static_cast<uint8_t*>(malloc(data.size()));
+        if (objectData == nullptr)
+        {
+            return false;
+        }
+        std::copy(std::begin(data), std::end(data), objectData);
+
+        auto* obj = reinterpret_cast<Object*>(objectData);
+        if (!callObjectValidate(objectHeader.getType(), *obj))
+        {
+            return false;
+        }
+
+        if (_totalNumImages >= Gfx::G1ExpectedCount::kObjects + Gfx::G1ExpectedCount::kDisc)
+        {
+            // Free objectData?
+            return false;
+        }
+
+        // Warning this saves a copy of the objectData pointer and must be unloaded prior to exiting this function
+        if (!partialLoad(objectHeader, stdx::span(objectData, data.size())))
+        {
+            return false;
+        }
+
+        // Object already installed so no need to install it
+        if (isObjectInstalled(objectHeader))
+        {
+            unloadAll();
+            return false;
+        }
+
+        bool result = installObject(objectHeader);
+
+        unloadAll();
+        return result;
+    }
+
+    // 0x00472AFE
+    ObjIndexPair getActiveObject(ObjectType objectType, uint8_t* edi)
+    {
+        const auto objects = getAvailableObjects(objectType);
+
+        for (auto [index, object] : objects)
+        {
+            if (edi[index] & (1 << 0))
+            {
+                return { static_cast<int16_t>(index), object };
+            }
+        }
+
+        return { -1, ObjectIndexEntry{} };
+    }
+}
